Extract createImagesInfiniteOptions into its own file
Pure mechanical move of the createImagesInfiniteOptions helper out of
useImagesInfinite.ts into createImagesInfiniteOptions.ts. The hook itself
is now a thin wrapper that just wires the options into createInfiniteQuery
and exposes refresh.
